barechatinput: moving helpers to shared package and adding test
