+++ packages/shared/src/logic/index.ts
@@ -11,3 +11,4 @@ export * from './tiptap';
 export * from './hosting';
 export * from './noun';
 export * from './wayfinding';
+export * from './content-helpers';
